Create codeql-analysis.yml (#2)
I haven't used this before, but we will be dealing with people's personal information at some point, so it's a good time to figure out how this works and if it's right for us.  The check came through fine, my only concern is that I may not have set it up correctly.